Still trying to figure out how to handle the duality of a component
It contains props which should be handled by both front-end and
back-end, and implementation, which should be handled only by
back-end.

We need a way for the front-end to import a component without
importing the implementation part.

Maybe parcel's tree-shaking is a good place to look for help? But
something tells me there's a better, more native way to handle that...

Another option would consist of two steps:

1. keeping props and components as separate imports
2. Not importing either during startup, requirering to import them explicitly